fix: do not create MFDs for empty region
